# java/src/main/resources/openapi-chat.yaml
openapi: 3.0.3
info:
  title: Netdisk API - 聊天功能
  description: 网盘系统聊天功能接口文档
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: 本地开发环境
paths:
...
  /api/v2/chat/conversations/{conversationId}/members:
    get:
      tags:
        - 成员管理
      summary: 获取成员列表
      description: 获取会话的成员列表
      operationId: listMembers
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberListResponse'

    post:
      tags:
        - 成员管理
      summary: 添加成员
      description: 向群聊添加成员
      operationId: addMembers
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMembersRequest'
      responses:
        '200':
          description: 添加成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddMembersResponse'
  /api/v2/chat/conversations/{conversationId}/members/{userId}:
    delete:
      tags:
        - 成员管理
      summary: 移除成员
      description: 从群聊中移除指定成员（仅群主/管理员可执行）
      operationId: removeMember
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
        - name: userId
          in: path
          required: true
          description: 被移除的用户UUID
          schema:
            type: string
      responses:
        '200':
          description: 移除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoveMemberResponse'
...
  /api/v2/chat/conversations/{conversationId}/invites:
    post:
      tags:
        - 群聊邀请
      summary: 创建邀请
      description: 群主或管理员邀请用户加入群聊
      operationId: createInvite
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInviteRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateInviteResponse'
        '400':
          description: 参数错误
        '403':
          description: 无权限

    get:
      tags:
        - 群聊邀请
      summary: 获取邀请列表
      description: 获取群聊的邀请记录列表
      operationId: listInvites
      security:
        - BearerAuth: []
      parameters:
        - name: conversationId
          in: path
          required: true
          description: 会话UUID
          schema:
            type: string
        - name: status
          in: query
          description: 筛选状态（pending/accepted/rejected/expired/cancelled）
          schema:
            type: string
        - name: page
          in: query
          description: 页码
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: 每页条数
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteListResponse'
  /api/v2/chat/invites/{inviteId}:
    get:
      tags:
        - 群聊邀请
      summary: 获取邀请详情
      description: 获取指定邀请的详细信息
      operationId: getInvite
      security:
        - BearerAuth: []
      parameters:
        - name: inviteId
          in: path
          required: true
          description: 邀请UUID
          schema:
            type: string
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InviteDetailResponse'

    delete:
      tags:
        - 群聊邀请
      summary: 取消邀请
      description: 取消一个待处理的邀请（仅邀请人可操作）
      operationId: cancelInvite
      security:
        - BearerAuth: []
      parameters:
        - name: inviteId
          in: path
          required: true
          description: 邀请UUID
          schema:
            type: string
      responses:
        '200':
          description: 取消成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelInviteResponse'
  /api/v2/chat/invites/{inviteToken}/accept:
    post:
      tags:
        - 群聊邀请
      summary: 接受邀请
      description: 通过邀请Token接受邀请加入群聊
      operationId: acceptInvite
      security:
        - BearerAuth: []
      parameters:
        - name: inviteToken
          in: path
          required: true
          description: 邀请Token
          schema:
            type: string
      responses:
        '200':
          description: 接受成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptInviteResponse'
        '400':
          description: 邀请已失效或已过期
        '403':
          description: 无权接受此邀请
  /api/v2/chat/invites/{inviteId}/reject:
    post:
      tags:
        - 群聊邀请
      summary: 拒绝邀请
      description: 拒绝一个收到的邀请
      operationId: rejectInvite
      security:
        - BearerAuth: []
      parameters:
        - name: inviteId
          in: path
          required: true
          description: 邀请UUID
          schema:
            type: string
      responses:
        '200':
          description: 拒绝成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RejectInviteResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AddMembersRequest:
      type: object
      required:
        - userIds
      properties:
        userIds:
          type: array
          description: 要添加的用户UUID列表
          items:
            type: string
          example: ["user_uuid_3", "user_uuid_4"]
...
    MemberInfo:
      type: object
      properties:
        userId:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
        role:
          type: string
        joinedAt:
          type: string
    MemberListResponse:
      type: object
      properties:
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/MemberInfo'
...
    AddMembersResponse:
      type: object
      properties:
        conversationId:
          type: string
        addedCount:
          type: integer
        failedUsers:
          type: array
          items:
            type: string
          description: 添加失败的成员列表及原因
    RemoveMemberResponse:
      type: object
      properties:
        conversationId:
          type: string
        removedUserId:
          type: string
        removed:
          type: boolean
...
    CreateInviteRequest:
      type: object
      description: 创建邀请请求
      properties:
        userId:
          type: string
          description: 被邀请用户UUID（与email/phone二选一）
          example: "user_uuid_xxx"
        email:
          type: string
          description: 被邀请人邮箱（与userId/phone二选一）
          example: "invitee@example.com"
        phone:
          type: string
          description: 被邀请人手机号（与userId/email二选一）
          example: "13800138000"
    CreateInviteResponse:
      type: object
      properties:
        inviteId:
          type: string
          description: 邀请UUID
          example: "invite_uuid_xxx"
        inviteToken:
          type: string
          description: 邀请Token（用于分享）
          example: "abc123def456"
        expiresAt:
          type: string
          format: date-time
          description: 过期时间
        inviteeUserId:
          type: string
          description: 被邀请用户UUID
        inviteeEmail:
          type: string
          description: 被邀请人邮箱
        inviteePhone:
          type: string
          description: 被邀请人手机号
    InviteListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/InviteDetailResponse'
        page:
          type: integer
          description: 当前页码
        pageSize:
          type: integer
          description: 每页条数
    InviteDetailResponse:
      type: object
      properties:
        id:
          type: integer
          description: 邀请记录ID
        inviteUuid:
          type: string
          description: 邀请UUID
        conversationId:
          type: string
          description: 会话UUID
        inviterUserId:
          type: string
          description: 邀请人用户UUID
        inviterNickname:
          type: string
          description: 邀请人昵称
        inviteeUserId:
          type: string
          description: 被邀请用户UUID
        inviteeEmail:
          type: string
          description: 被邀请人邮箱
        inviteePhone:
          type: string
          description: 被邀请人手机号
        inviteToken:
          type: string
          description: 邀请Token
        status:
          type: string
          enum: [pending, accepted, rejected, expired, cancelled]
          description: 邀请状态
        expiresAt:
          type: string
          format: date-time
          description: 过期时间
        acceptedAt:
          type: string
          format: date-time
          description: 接受时间
        rejectedAt:
          type: string
          format: date-time
          description: 拒绝时间
        createdAt:
          type: string
          format: date-time
          description: 创建时间
...
    AcceptInviteResponse:
      type: object
      properties:
        accepted:
          type: boolean
          description: 是否接受成功
        conversationId:
          type: string
          description: 会话UUID
    RejectInviteResponse:
      type: object
      properties:
        rejected:
          type: boolean
          description: 是否拒绝成功
    CancelInviteResponse:
      type: object
      properties:
        cancelled:
          type: boolean
          description: 是否取消成功